specializations/aegis/seed/requirements.txt
torch>=2.1.0
numpy>=1.24.0
einops>=0.7.0

# Database and storage
pymongo>=4.5.0
redis>=5.0.0
psycopg2-binary>=2.9.0

# Distributed computing
ray>=2.7.0

# Monitoring and logging
wandb>=0.15.0
tensorboard>=2.14.0

# Configuration
hydra-core>=1.3.0
pydantic>=2.4.0
omegaconf>=2.3.0

# API and web
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# Scientific computing
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
jsonschema>=4.19.0

# Code generation and analysis (optional)
openai>=1.0.0  # For LLM-based code generation
anthropic>=0.7.0  # For Claude API

# Safety and testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.88.0  # Property-based testing

# Acceleration
accelerate>=0.24.0
safetensors>=0.4.0

# Development tools
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
